First initialization
This is my first initialization. This is the starting stage in which I made. So to this point, I've only made the login page and the main dashboard. The main dashboard shows the beds in which we can admit patients to. For now, the moment you click that button, it'll show the previously hidden admission forms. In the next update, I'll separate the admission forms into a new page.